src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.48
botocore==1.34.48
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
exif==1.6.0
expiringdict==1.2.2
fastapi==0.110.0
ffmpeg-python==0.2.0
filetype==1.2.0
future==1.0.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
mangum==0.17.0
pillow==10.2.0
pkgconfig==1.5.5
plum-py==0.8.7
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-magic==0.4.27
pyvips==2.2.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.27.1
